// apps/web/src/contracts/generated/sheet/MapLayers.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Empty } from "./Empty";
import type { LayerList } from "./LayerList";
import type { ThemeRef } from "./ThemeRef";

/**
 * Which layers a map shows.
 */
export type MapLayers = { "type": "all" } & Empty | { "type": "theme" } & ThemeRef | { "type": "list" } & LayerList;